Применил в ленте новостей паттерны push-to-refresh, infinite scrolling и расчет ячеек для фотографий

+++ VK_App/JSON Mapping/NewsData.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct NewsData {
     let sourceID: Int
-    let date: Int
+    let date: Double
     let text: String?
     let comments: NewsComments
     let likes: NewsLikes
